refactor: ダミーデータファイルの場所を変更
コンポーネントと分離するため
